Write test sqlite vector db

+++ tests/vectordb/test_sqlite_vectordb.py
@@ -2,14 +2,15 @@ import pytest
 from nchain.vectordb.sqlite_vectordb import SQLiteVectorDB
 import os
 
-DB_PATH = "tests/resources/test_sql_vector.db"
+DB_PATH = "tests/resources/vector.db"
+INDEXDB_PATH = "tests/resources/index.ann"
 
 #@pytest.fixture(scope="module")
 @pytest.fixture(scope="function")
 def db():
     if os.path.exists(DB_PATH):
         os.remove(DB_PATH)
-    database = SQLiteVectorDB(dimension=2, db_path=DB_PATH)
+    database = SQLiteVectorDB(dimension=2, indexdb_path=INDEXDB_PATH, db_path=DB_PATH)
     #yield database
     #os.remove(DB_PATH)
     return database
@@ -18,19 +19,22 @@ def test_vector_storage_and_search(db:SQLiteVectorDB):
     # Storing a single vector
     vector = [1.0, 2.0]
     metadata = {"info": "sample_vector"}
-    db.store_vectors([vector], [metadata])
+    db.store_vectors([vector], [metadata], [b"sample_content"], True)
 
+    other_vector = [3.0, 3.0]
+    other_metadata = {"info": "sample_vector 2"}
+    db.store_vectors([other_vector], [other_metadata], [b"sample_content_other"], True)
     # Searching the vector itself should return the same vector as the most similar one
     results = db.search_vectors(vector, top_k=1)
     assert len(results) == 1
-    _, _, meta = results[0]
-    assert meta["info"] == "sample_vector"
-
+    entry = results[0]
+    assert entry.metadata["info"] == "sample_vector"
+"""
 def test_vector_update(db:SQLiteVectorDB):
     # Storing a single vector
     vector = [1.0, 2.0]
     metadata = {"info": "sample_vector"}
-    db.store_vectors([vector], [metadata])
+    db.store_vectors([vector], [metadata], [b"sample_content"], True)
 
     idx, _, _ = db.search_vectors(vector, top_k=1)[0]
 
@@ -50,7 +54,7 @@ def test_vector_deletion():
     # Storing two vectors
     vectors = [[1.0, 2.0], [2.0, 3.0]]
     metadata_list = [{"info": "first_vector"}, {"info": "second_vector"}]
-    db.store_vectors(vectors, metadata_list)
+    db.store_vectors(vectors, metadata_list, [b"sample_content"] * len(vectors), True)
 
     idx0, disntace0, meta0 = db.search_vectors(vectors[0], top_k=1)[0]
     # Delete the first vector
@@ -61,7 +65,9 @@ def test_vector_deletion():
     assert len(results) == 1
     _, _, meta = results[0]
     assert meta["info"] == "second_vector"
+"""
 
 #def teardown_module():
 def teardown_function():
-    os.remove(DB_PATH)
+    if os.path.exists(DB_PATH):
+        os.remove(DB_PATH)
